+++ src/move-nodes.ts
@@ -44,10 +44,27 @@ export function isFuture(inputNode: TaskPaperNode): boolean {
     );
 }
 
-export function isDueToday(inputNode: TaskPaperNode): boolean {
+export function isDueTodayOrBefore(inputNode: TaskPaperNode): boolean {
     return (
         inputNode.hasTag("due") &&
         !inputNode.hasTag("done") &&
         cleanDate(inputNode.tagValue("due")).isSameOrBefore(todayDay, "day")
+    );
+}
+
+export function isDueToday(inputNode: TaskPaperNode): boolean {
+    return (
+        inputNode.hasTag("due") &&
+        !inputNode.hasTag("done") &&
+        cleanDate(inputNode.tagValue("due")).isSame(todayDay, "day")
+    );
+}
+
+
+export function isOverdue(inputNode: TaskPaperNode): boolean {
+    return (
+        inputNode.hasTag("due") &&
+        !inputNode.hasTag("done") &&
+        cleanDate(inputNode.tagValue("due")).isBefore(todayDay, "day")
     );
 }